+++ binary_cnn/CNN/conv_layer_bin_t.h
@@ -5,20 +5,20 @@
 struct conv_layer_bin_t
 {
 	layer_type type = layer_type::conv_bin;
-	tensor_t<double> grads_in;
-	tensor_t<double> in;
-	tensor_t<double> out;
+	tensor_t<float> grads_in;
+	tensor_t<float> in;
+	tensor_t<float> out;
 	tensor_bin_t in_bin; 		// 1st BINARIZATION (h1)
 	tensor_bin_t in_bin2;		// 2nd BINARIZATION (h2)
-	tensor_t<double> al_b; 			// α1 * h1 + α2 * h2
+	tensor_t<float> al_b; 			// α1 * h1 + α2 * h2
 	tensor_bin_t out_bin;
-	tensor_t<double> filters; 
+	tensor_t<float> filters; 
 	tensor_bin_t filters_bin; 
 	tensor_t<gradient_t> filter_grads;
 	uint16_t stride;
 	uint16_t extend_filter;
-	vector<double> alpha;
-	vector<double> alpha2;
+	vector<float> alpha;
+	vector<float> alpha2;
 	bool debug,clip_gradients_flag; 	
 	conv_layer_bin_t( uint16_t stride, uint16_t extend_filter, uint16_t number_filters, tdsize in_size ,
 	bool clip_gradients_flag = true, bool debug_flag = false)
@@ -47,11 +47,11 @@ struct conv_layer_bin_t
 		this->stride = stride;
 		this->extend_filter = extend_filter;
 		this->clip_gradients_flag = clip_gradients_flag;
-		assert( (double( in_size.x - extend_filter ) / stride + 1)
+		assert( (float( in_size.x - extend_filter ) / stride + 1)
 				==
 				((in_size.x - extend_filter) / stride + 1) );
 
-		assert( (double( in_size.y - extend_filter ) / stride + 1)
+		assert( (float( in_size.y - extend_filter ) / stride + 1)
 				==
 				((in_size.y - extend_filter) / stride + 1) );
 
@@ -61,7 +61,7 @@ struct conv_layer_bin_t
 				for ( int j = 0; j < extend_filter; j++ ){
 					for ( int z = 0; z < in_size.z; z++ ){
 						 //initialization of floating weights 
-						filters(a,i, j, z ) =  (1.0f * (rand()-rand())) / double( RAND_MAX );
+						filters(a,i, j, z ) =  (1.0f * (rand()-rand())) / float( RAND_MAX );
 
 						// initialization of binary weights
 						filters_bin.data[filters_bin(a,i,j,z)] = 0;
@@ -92,7 +92,7 @@ struct conv_layer_bin_t
 		int max_x, max_y, max_z;
 	};
 
-	int normalize_range( double f, int max, bool lim_min )
+	int normalize_range( float f, int max, bool lim_min )
 	{
 		if ( f <= 0 )
 			return 0;
@@ -108,8 +108,8 @@ struct conv_layer_bin_t
 
 	range_t map_to_output( int x, int y )
 	{
-		double a = x;
-		double b = y;
+		float a = x;
+		float b = y;
 		return
 		{
 			normalize_range( (a - extend_filter + 1) / stride, out.size.x, true ),
@@ -121,7 +121,7 @@ struct conv_layer_bin_t
 		};
 	}
 
-	void activate( tensor_t<double>& in )
+	void activate( tensor_t<float>& in )
 	{
 		this->in = in;
 		activate();
@@ -174,7 +174,7 @@ struct conv_layer_bin_t
 
 		// CALCULATE alpha1		
 		for(int e = 0; e<in.size.m; e++){
-			double sum = 0;
+			float sum = 0;
 			for(int x=0; x<in.size.x; x++)
 				for(int y=0; y<in.size.y; y++)
 					for(int z=0; z<in.size.z; z++){
@@ -189,15 +189,15 @@ struct conv_layer_bin_t
 		}
 
 		// CALCULATE alpha2
-		tensor_t<double> temp(in.size.m, in.size.x, in.size.y, in.size.z);
+		tensor_t<float> temp(in.size.m, in.size.x, in.size.y, in.size.z);
 		for(int e = 0; e<in.size.m; e++){
 			
-			double sum = 0;
+			float sum = 0;
 			
 			for(int x=0; x<in.size.x; x++)
 				for(int y=0; y<in.size.y; y++)
 					for(int z=0; z<in.size.z; z++){
-						temp(e,x,y,z) = in(e,x,y,z) - alpha[e]*(in_bin(e,x,y,z)==1 ? double(1) : double(-1) );
+						temp(e,x,y,z) = in(e,x,y,z) - alpha[e]*(in_bin(e,x,y,z)==1 ? float(1) : float(-1) );
 						in_bin2.data[in_bin2(e,x,y,z)] = temp(e,x,y,z)>=0? 1 : 0;
 						sum += abs(temp(e,x,y,z));
 				}
@@ -222,8 +222,8 @@ struct conv_layer_bin_t
 			for(int x=0; x<in.size.x; x++)
 				for(int y=0; y<in.size.y; y++)
 					for(int z=0; z<in.size.z; z++){
-						al_b(e, x, y, z) = alpha[e] * (in_bin(e, x, y, z) == 1 ? double(1) : double(-1) ) +
-								alpha2[e] * (in_bin2(e, x, y, z) == 1 ? double(1) : double(-1) );
+						al_b(e, x, y, z) = alpha[e] * (in_bin(e, x, y, z) == 1 ? float(1) : float(-1) ) +
+								alpha2[e] * (in_bin2(e, x, y, z) == 1 ? float(1) : float(-1) );
 					}
 		
 		if(debug)
@@ -250,7 +250,7 @@ struct conv_layer_bin_t
 					for ( int y = 0; y < out.size.y; y++ ){
 
 						point_t mapped = map_to_input( { 0, (uint16_t)x, (uint16_t)y, 0 }, 0 );
-						double sum = 0, sum2 = 0;
+						float sum = 0, sum2 = 0;
 						for ( int i = 0; i < extend_filter; i++ )
 							for ( int j = 0; j < extend_filter; j++ )
 								for ( int z = 0; z < in.size.z; z++ )
@@ -280,13 +280,13 @@ struct conv_layer_bin_t
 	
 	
 	
-	void fix_weights(double learning_rate){
+	void fix_weights(float learning_rate){
 		for ( int a = 0; a < filters.size.m; a++ )
 			for ( int i = 0; i < extend_filter; i++ )
 				for ( int j = 0; j < extend_filter; j++ )
 					for ( int z = 0; z < in.size.z; z++ )
 					{
-						double& w = filters(a, i, j, z );
+						float& w = filters(a, i, j, z );
 						gradient_t& grad = filter_grads(a, i, j, z );
 						grad.grad /= in.size.m;
 						w = update_weight( w, grad, 1, true, learning_rate);
@@ -299,7 +299,7 @@ struct conv_layer_bin_t
 		}
 	}
 
-	void calc_grads( tensor_t<double>& grad_next_layer)
+	void calc_grads( tensor_t<float>& grad_next_layer)
 	{
 		for ( int k = 0; k < filter_grads.size.m; k++ )
 		{
@@ -315,7 +315,7 @@ struct conv_layer_bin_t
 					range_t rn = map_to_output( x, y );
 					for ( int z = 0; z < in.size.z; z++ ){
 						
-						double sum_error = 0;
+						float sum_error = 0;
 						for ( int i = rn.min_x; i <= rn.max_x; i++ ){
 							int minx = i * stride;
 							for ( int j = rn.min_y; j <= rn.max_y; j++ ){
@@ -325,7 +325,7 @@ struct conv_layer_bin_t
 									filter_grads(k, x - minx, y - miny, z ).grad += al_b(e, x, y, z ) * grad_next_layer(e, i, j, k );
 									clip_gradients(clip_gradients_flag, filter_grads(k, x - minx, y - miny, z ).grad);
 									if(in(e,x,y,z) <= 1){
-										double w_applied = (filters_bin.data[filters_bin(k, x - minx, y - miny, z )] == 1? double(1) : double(-1));
+										float w_applied = (filters_bin.data[filters_bin(k, x - minx, y - miny, z )] == 1? float(1) : float(-1));
 										sum_error += w_applied * grad_next_layer(e, i, j, k );
 									}
 									else sum_error = 0;
