+++ binary_cnn/Example MNIST/full_test/training_test_mnist_binary1.cpp
@@ -14,8 +14,8 @@ int num_batches = 6000/mini_batch_size;
 	
 struct case_t
 {
-	tensor_t<double> data;
-	tensor_t<double> out;
+	tensor_t<float> data;
+	tensor_t<float> out;
 };
 
 uint8_t* read_file( const char* szFile )
@@ -51,7 +51,7 @@ vector<case_t> read_test_cases()
 	
     int crnt_count = 0;
 
-	case_t c {tensor_t<double>( mini_batch_size, 28, 28, 1 ), tensor_t<double>(mini_batch_size, 10, 1, 1 )};
+	case_t c {tensor_t<float>( mini_batch_size, 28, 28, 1 ), tensor_t<float>(mini_batch_size, 10, 1, 1 )};
  
     for( int i = 0; i < case_count; i++ )
 	{
@@ -125,14 +125,14 @@ int main()
                 // print_tensor(layer5->out);
                 // }
                 
-                // tensor_t<double> costs = cross_entropy(layer6->out, cases[batch_num].out);
-                // double costs_avg = 0;
+                // tensor_t<float> costs = cross_entropy(layer6->out, cases[batch_num].out);
+                // float costs_avg = 0;
 
                 // for(int e = 0; e<mini_batch_size; e++){
                 //     costs_avg += costs(e,0,0,0);
                 // }
 
-                // costs_avg /= (double)mini_batch_size;
+                // costs_avg /= (float)mini_batch_size;
                 
                 // cost_vec.push_back(costs_avg);
                 
@@ -165,7 +165,7 @@ int main()
                 layer2->calc_grads(layer3->grads_in);
                 layer1->calc_grads(layer2->grads_in);
                 
-                // double diff_cost = cost_vec[cost_vec.size()-1] - cost_vec[cost_vec.size()-2];
+                // float diff_cost = cost_vec[cost_vec.size()-1] - cost_vec[cost_vec.size()-2];
                 // if ( diff_cost < 0.00001 )     // Stops training if cost decreases very slow
                 //   break;
                 // cout<<"*************epoch number*********** "<<epoch<<"***********************\n";
